New - dovecot and local smtp-auth
This represents another major step towards configuring astro as our new mail
server.  Dovecot is now installed and configured to service IMAP connections
and also to provide a local socket for Postfix to utilize for SMTP-AUTH.  The
OpenSSL module now allows certificate installation w/o a corresponding key,
such as is required with a CA certificate.

- install and configure Dovecot
- configure Postfix to use Dovecot's socket for SMTP-AUTH
- install doubledog-cacert certificate
- install doubledog-astro-imap certificate
- install doubledog-astro-smtp certificate
- fix postfix class to use "modules" in file source reference