Doc - make clear that notify is permitted
Resources inside a define inherit any metaparameters passed into the define.
For my openssl::tls-certificate, this can be useful to notify a service of any
changes to the private key or the public certificate.  This is useful since I
do not know of anyway for the service to subscribe directly to those files
within the define.