Change - allow override of owner/group for TLS certificate files
In some instances, such as with OpenLDAP, it is necessary for the
certificate and key files to be owned by less than root.  The slapd
daemon runs as the ldap user and needs read access to the private key
(and certificate) yet the permissions cannot be weakened; this is the
correct approach.