Drop - openssl.cnf management
This essentially did nothing but set defaults to make certificate
creation via the openssl CLI a bit easier.  However, that same CLI sucks
when it comes to handling v3 extensions like SubjectAltNames as a custom
config is needed to do those.  I was not able to make work any of the
tricks to obtain values via shell variables, which was hackish at best.
A much better approach is to simply use an awesome tool like xca.  With
that, there's no point in waiting for puppet to manage this for no gain.